check_miami_input {miamiplot}	R Documentation

Check input to miami plot

Description

Check input to miami plot

Usage

check_miami_input(data, split_by, split_at, chr = "chr", pos = "pos", p = "p")

Arguments

data
A data.frame object. Required.

split_by
A character vector. The name of the column to use for splitting into upper and lower sections of the Miami plot. Required.

split_at
A character or numeric vector. If numeric, the upper plot will contain your results where the values in the split_by column are >= split_at. If character, upper plot will contain your results where the values in the split_by column are equal to split_at. Required.

chr
The name of the column containing your chromosome information. Defaults to "chr"

pos
The name of the column containing your position information. Defaults to "pos"

p
The name of the column containing your p-value information. Defaults to "p"

Value

If no problems, executes silently. Will return an error message if there are problems.

Author(s)

Julie White

References

https://github.com/juliedwhite/miamiplot

Examples

check_miami_input(data = gwas_results, split_by = "beta",
                            split_at = 0, p = "pval")

[Package miamiplot version 1.1.0 Index]