gwas_results {miamiplot}	R Documentation

Simulated data for 60,000 SNPs from two studies

Description

A simulated dataset containing the rsid, chromosome, position, beta value, standard error, t-statistic, p-value, and study name for 60,000 SNPs total. Each study has 30,000 SNPs spread across 22 chromosomes.

Usage

gwas_results

Format

A data.frame with 60,000 rows and 8 variables:

rsid

Variant name

chr

Chromosome

pos

Base pair position

beta

Beta value from simulated GWAS

se

Standard error from simulated GWAS

tstat

T-statistic from simulated GWAS

pval

P-value from simulated GWAS

study

Either study A or study B, to signify two GWAS of the same phenotypes that have been combined for plotting purposes.

[Package miamiplot version 1.1.0 Index]